display MI3622 report after calculation
